# server/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ContentPermissions:
    properties:
      comment:
        type: boolean
      delete:
        type: boolean
      edit:
        type: boolean
      view:
        type: boolean
      viewBody:
        type: boolean
      vote:
        type: boolean
    required:
    - comment
    - delete
    - edit
    - view
    - viewBody
    - vote
    type: object
...
  dto.EditorialCreateRequest:
    properties:
      contentMd:
        type: string
      problemId:
        type: string
      solvedOnly:
        type: boolean
      status:
        enum:
        - draft
        - published
        type: string
      title:
        type: string
      visibility:
        enum:
        - public
        - private
        type: string
    required:
    - contentMd
    - problemId
    - title
    type: object
...
  dto.EditorialResponse:
    properties:
      authorId:
        type: string
      authorName:
        type: string
      canEdit:
        type: boolean
      contentMd:
        type: string
      createdAt:
        type: string
      domainId:
        type: string
      id:
        type: string
      locked:
        description: |-
          Locked is true when the body was withheld because the reader has not
          solved the problem yet.
        type: boolean
      permissions:
        $ref: '#/definitions/dto.ContentPermissions'
      problemId:
        type: string
      problemPublicId:
        type: string
      problemTitle:
        type: string
      publicId:
        type: string
      solvedOnly:
        type: boolean
      status:
        enum:
        - draft
        - published
        type: string
      title:
        type: string
      updatedAt:
        type: string
      visibility:
        enum:
        - public
        - private
        type: string
      voteCount:
        type: integer
      voted:
        type: boolean
    required:
    - canEdit
    - contentMd
    - createdAt
    - domainId
    - id
    - locked
    - permissions
    - problemId
    - problemPublicId
    - publicId
    - solvedOnly
    - status
    - title
    - updatedAt
    - visibility
    - voteCount
    - voted
    type: object
  dto.EditorialSummaryResponse:
    properties:
      authorId:
        type: string
      authorName:
        type: string
      canEdit:
        type: boolean
      createdAt:
        type: string
      domainId:
        type: string
      id:
        type: string
      locked:
        type: boolean
      permissions:
        $ref: '#/definitions/dto.ContentPermissions'
      problemId:
        type: string
      problemPublicId:
        type: string
      problemTitle:
        type: string
      publicId:
        type: string
      solvedOnly:
        type: boolean
      status:
        enum:
        - draft
        - published
        type: string
      title:
        type: string
      updatedAt:
        type: string
      visibility:
        enum:
        - public
        - private
        type: string
      voteCount:
        type: integer
      voted:
        type: boolean
    required:
    - canEdit
    - createdAt
    - domainId
    - id
    - locked
    - permissions
    - problemId
    - problemPublicId
    - publicId
    - solvedOnly
    - status
    - title
    - updatedAt
    - visibility
    - voteCount
    - voted
    type: object
  dto.EditorialUpdateRequest:
    properties:
      contentMd:
        type: string
      solvedOnly:
        type: boolean
      status:
        enum:
        - draft
        - published
        type: string
      title:
        type: string
      visibility:
        enum:
        - public
        - private
        type: string
    required:
    - contentMd
    - title
    type: object
  dto.EditorialVoteRequest:
    properties:
      up:
        description: Up records an upvote; false withdraws it.
        type: boolean
    required:
    - up
    type: object
  dto.EditorialVoteResponse:
    properties:
      voteCount:
        type: integer
      voted:
        type: boolean
    required:
    - voteCount
    - voted
    type: object
...
  httpx.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    required:
    - code
    - error
    type: object
...
  httpx.ListResponse-dto_EditorialSummaryResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.EditorialSummaryResponse'
        type: array
      total:
        type: integer
    required:
    - items
    - total
    type: object
...
  httpx.StatusResponse:
    properties:
      status:
        type: string
    required:
    - status
    type: object
info:
  contact: {}
  description: Public application endpoints and the authenticated Judge worker protocol.
  title: Vertex OJ API
  version: "1.0"
paths:
...
  /api/editorials:
    get:
      parameters:
      - description: Problem ID
        in: query
        name: problem
        type: string
      - description: Author user ID
        in: query
        name: author
        type: string
      - description: Search text
        in: query
        name: keyword
        type: string
      - description: Ordering
        enum:
        - recent
        - votes
        in: query
        name: sort
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.ListResponse-dto_EditorialSummaryResponse'
      summary: List editorials
      tags:
      - editorials
    post:
      consumes:
      - application/json
      parameters:
      - description: Editorial
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorialCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EditorialResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an editorial
      tags:
      - editorials
  /api/editorials/{id}:
    delete:
      parameters:
      - description: Editorial ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.StatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an editorial
      tags:
      - editorials
    get:
      parameters:
      - description: Editorial ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorialResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get an editorial
      tags:
      - editorials
    put:
      consumes:
      - application/json
      parameters:
      - description: Editorial ID
        in: path
        name: id
        required: true
        type: string
      - description: Editorial
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorialUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorialResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an editorial
      tags:
      - editorials
...
  /api/editorials/{id}/vote:
    post:
      consumes:
      - application/json
      parameters:
      - description: Editorial ID
        in: path
        name: id
        required: true
        type: string
      - description: Vote
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorialVoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorialVoteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Vote for an editorial
      tags:
      - editorials
...
securityDefinitions:
  BearerAuth:
    description: Prefix the access token with "Bearer ".
    in: header
    name: Authorization
    type: apiKey
  JudgeServiceAuth:
    description: Internal Judge service credential using the Bearer scheme.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"